changes in application.properties and build.gradle

+++ demo/demo/src/main/java/com/example/demo/dto/DogShowMustacheDto.java
@@ -14,8 +14,8 @@ public class DogShowMustacheDto {
     Long id;
     String dogsOwnerName;
     String name;
-    Integer age;
-    Long weight;
+    Double age;
+    Double weight;
     String sex;
     static OwnerService ownerService;
